+++ rust/main/submitter/src/chain_tx_adapter.rs
@@ -1,8 +1,10 @@
 // TODO: re-enable clippy warnings
 #![allow(unused_imports)]
 
+pub use adapter::{AdaptsChain, GasLimit};
+pub use chains::ChainTxAdapterFactory;
+pub use chains::SealevelPayload;
+pub use chains::SealevelTxPrecursor;
+
 mod adapter;
 mod chains;
-
-pub use adapter::*;
-pub use chains::*;
