+++ rust/main/submitter/src/chain_tx_adapter/chains.rs
@@ -1,6 +1,6 @@
-pub(crate) use factory::ChainTxAdapterFactory;
-pub(crate) use sealevel::SealevelPayload;
-pub(crate) use sealevel::SealevelTxPrecursor;
+pub use factory::ChainTxAdapterFactory;
+pub use sealevel::SealevelPayload;
+pub use sealevel::SealevelTxPrecursor;
 
 mod factory;
