+++ rust/main/submitter/src/chain_tx_adapter/chains/sealevel.rs
@@ -1,6 +1,6 @@
-pub(crate) use adapter::SealevelTxAdapter;
-pub(crate) use payload::SealevelPayload;
-pub(crate) use precursor::SealevelTxPrecursor;
+pub use adapter::SealevelTxAdapter;
+pub use payload::SealevelPayload;
+pub use precursor::SealevelTxPrecursor;
 
 mod adapter;
 mod conf;
