+++ rust/main/submitter/src/chain_tx_adapter/chains/sealevel/precursor.rs
@@ -7,20 +7,20 @@ use crate::chain_tx_adapter::chains::sealevel::payload::Instruction;
 use crate::payload::FullPayload;
 
 #[derive(Debug, Clone, serde::Deserialize, serde::Serialize, PartialEq, Eq)]
-pub(crate) struct SealevelTxPrecursor {
-    pub(crate) instruction: SealevelInstruction,
-    pub(crate) estimate: SealevelTxCostEstimate,
+pub struct SealevelTxPrecursor {
+    pub instruction: SealevelInstruction,
+    pub estimate: SealevelTxCostEstimate,
 }
 
 impl SealevelTxPrecursor {
-    pub(crate) fn new(instruction: SealevelInstruction, estimate: SealevelTxCostEstimate) -> Self {
+    pub fn new(instruction: SealevelInstruction, estimate: SealevelTxCostEstimate) -> Self {
         Self {
             instruction,
             estimate,
         }
     }
 
-    pub(crate) fn from_payload(payload: &FullPayload) -> Self {
+    pub fn from_payload(payload: &FullPayload) -> Self {
         let instruction = payload.instruction();
         SealevelTxPrecursor::new(instruction.clone(), SealevelTxCostEstimate::default())
     }
