+++ rust/main/submitter/src/chain_tx_adapter/chains/sealevel/transaction.rs
@@ -1,5 +1,7 @@
 pub use factory::TransactionFactory;
 pub use precursor::Precursor;
+pub use update::Update;
 
 mod factory;
 mod precursor;
+mod update;
